tools: Make clang-format-diff.py use versioned files, not local files
clang-format-diff.py uses the 'diff' passed in stdin to know which files
to pass to clang-format. clang-format then uses the *local* versions of those
files to check any formatting issues: then if formatting is correct locally
but not on HEAD (which is currently being pushed), the script does not
complain and then we can push unformatted source files.

Fix that by checkouting those files into /tmp directory and make clang-format
work on the temp files.